refactored and moved helpers to separate file
